pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tab-audit"
version = "0.2.0"
description = "Config-driven tabular dataset quality auditor for OpenML, Kaggle, and local files"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Tab Audit"}]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pandas>=2.1.0",
  "pyarrow>=14.0.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.0",
  "jinja2>=3.1.0",
  "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
openml = ["openml>=0.14.2"]
kaggle = ["kaggle>=1.6.17"]
profiling = ["ydata-profiling>=4.9.0"]
ml = ["pandera>=0.20.4", "deepchecks>=0.18.1", "cleanlab>=2.7.0"]
gpu-xgboost = ["xgboost>=2.1.0"]
gpu-lightgbm = ["lightgbm>=4.5.0"]
gpu-catboost = ["catboost>=1.2.7"]
gpu = ["xgboost>=2.1.0", "lightgbm>=4.5.0", "catboost>=1.2.7"]
test = ["pytest>=8.0.0", "pytest-cov>=5.0.0"]
lint = ["ruff>=0.6.9", "pylint>=3.2.7"]

[project.scripts]
tabaudit = "tab_audit.cli:main"

[tool.setuptools]
packages = [
  "tab_audit",
  "tab_audit.checks",
  "tab_audit.connectors",
  "tab_audit.io",
  "tab_audit.reporting",
  "tab_audit.scoring",
  "tab_audit.utils",
  "tab_audit.modeling",
  "tab_audit.leaderboard",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["tab_audit"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 60

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
